feat: chat box icon in profile dropdown. Should open up a chat room on right of page. Work in progress.
