+++ src/styles/main.css
@@ -238,6 +238,9 @@ nav button:hover {
     margin-top: 100px;
     background-color: aliceblue;
 }
+ .shrink {
+     width: 65% !important;
+ }
 
 #writeMessage {
     width: 300px;
